Transform app module to modern class with config parameter.
