+++ src/app/Container.jsx
@@ -1,0 +1,7 @@
+const Container = ({children}) => {
+  return (
+    <div className="container" style="background-color: green; height: 450px">{children}</div>
+  )
+}
+
+export default Container;
