--- src/routes.js
@@ -1,6 +1,0 @@
-/** @type {RouteDefinition[]} */
-const routes = [
-
-];
-
-export default routes;
